+++ src/App.css
@@ -1,5 +1,16 @@
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
+
+html {
+  padding: none;
+}
+
 body {
   min-width: 100vw;
+  margin: 0;
 }
 
 #login {
