+++ src/styles/layout/index.css
@@ -1,0 +1,27 @@
+#layout {
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+  min-width: 100vw;
+  position: relative;
+}
+
+#background {
+  background: url(/background-layout-Lims.png) no-repeat center center fixed;
+  background-size: cover;
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  z-index: -1;
+}
+
+#content {
+  display: flex;
+  color: blue;
+  max-width: 100vw;
+  max-height: 100vh;
+  box-sizing: border-box;
+  padding: 40px;
+}
