+++ src/styles/nav-lab-manager/index.css
@@ -1,0 +1,54 @@
+* {
+  box-sizing: border-box;
+}
+
+.nav-lab-header {
+  background-color: #0cb2aa;
+  color: #fff;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  gap: 2rem;
+  padding: 0 1rem;
+  align-items: center;
+  margin: 1rem;
+  border-radius: 5px;
+  max-width: 100vw;
+}
+
+.navbar-logo {
+  font-family: "DM Serif Text", serif;
+  font-size: 30px;
+  font-weight: bold;
+}
+
+.nav-links {
+  font-family: "Inter", sans-serif;
+  color: inherit;
+  font-size: 13px;
+  margin: 0;
+  padding: 0.4rem;
+  list-style: none;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 1.5rem;
+}
+
+a {
+  color: inherit;
+  text-decoration: none;
+}
+
+a:hover {
+  color: #200e32;
+  transition-duration: 150ms;
+}
+
+.custom-avatar {
+  background-color: #dce7f5;
+}
+
+.custom-avatar .anticon {
+  color: #0c1827;
+}
